OE-351 🔨 move tests to another namespace
